added getAllProductsFromOrder in service + tests

+++ src/main/java/com/ordersmicroservice/orders_microservice/models/OrderedProduct.java
@@ -24,6 +24,6 @@ public class OrderedProduct {
     Long product_id;
 
     @Column(name = "quantity")
-    Long quantity;
+    Integer quantity;
 }
